+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 <!-- The order of list items should be: Critical/Fixes, New, Update, Remove, Underpinnings -->
-<!-- ## [UNRELEASED](https://github.com/roydukkey/eslint-config/compare/v2.0.3...master) -->
+<!-- ## [UNRELEASED](https://github.com/roydukkey/eslint-config/compare/v2.0.4...master) -->
+
+## [2.0.4](https://github.com/roydukkey/eslint-config/compare/v2.0.3...v2.0.4)
+
+* Only apply `recommended-requiring-type-checking` to `*.ts?(x)` and `*.{m,c}ts`
+* Revert `sort-imports` member syntax sort order to v1 configuration
 
 ## [2.0.3](https://github.com/roydukkey/eslint-config/compare/v2.0.2...v2.0.3)